Fixing bug on 'very minor' input
When running the program with the "very minor" richter scale, it wouldn't fall into the "very minor" case since it only compared with os.Args[1] when in fact we need to concatenate os.Args[1] and os.Args[2] to retrieve the "very minor" command. This commit fixes this error by adding an additional if clause to check for this case.